Updating module with finished functionality.
