+++ docroot/modules/custom/prisoner_hub_prison_access_cms/prisoner_hub_prison_access_cms.services.yml
@@ -2,6 +2,6 @@ parameters:
   prisoner_hub_prison_access_cms.prison_owner_field_name: field_prison_owner
   prisoner_hub_prison_access_cms.user_prison_field_name: field_user_prisons
 services:
-  prisoner_hub_prison_access_cms.field_permissions:
-    class: Drupal\prisoner_hub_prison_access_cms\FieldPermissions
+  prisoner_hub_prison_access_cms.entity_edit_access:
+    class: Drupal\prisoner_hub_prison_access_cms\EntityEditAccess
     arguments: ['@current_user', '%prisoner_hub_prison_access_cms.prison_owner_field_name%', '%prisoner_hub_prison_access_cms.user_prison_field_name%', '%prisoner_hub_prison_access.exclude_from_prison_field_name%']
