+++ docroot/modules/custom/prisoner_hub_prison_access_cms/src/ExcludedFromFieldPreRender.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace Drupal\prisoner_hub_prison_access_cms;
+
+use Drupal\Core\Render\Element;
+use Drupal\Core\Security\TrustedCallbackInterface;
+
+/**
+ * Class ExcludedFromFieldPreRender
+ *
+ * Prerender callback for the field_exclude_from_prison.
+ */
+class ExcludedFromFieldPreRender implements TrustedCallbackInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function trustedCallbacks() {
+    return ['preRenderExcludeFromPrisonField'];
+  }
+
+  /**
+   * Set disabled states to prison taxonomy terms the user does not have accces to.
+   *
+   * @param array $element
+   *   The checkbox_tree element, generated by the term_reference_tree module.
+   *
+   * @return array
+   *   The modified array, with disabled states.
+   */
+  public static function preRenderExcludeFromPrisonField(array $element) {
+    foreach (Element::children($element) as $key) {
+      $entity_edit_access = \Drupal::service('prisoner_hub_prison_access_cms.entity_edit_access');
+      if (is_int($key) && !$entity_edit_access->hasPrisonTermAccess($key)) {
+        $element[$key]['#attributes']['disabled'] = 'disabled';
+      }
+      foreach (Element::children($element[$key]) as $child_key) {
+        $element[$key][$child_key] = self::preRenderExcludeFromPrisonField($element[$key][$child_key]);
+      }
+    }
+    return $element;
+  }
+}
